+++ packages/rasengan/src/routing/utils/flat-routes.tsx
@@ -1,29 +1,23 @@
 import { FunctionComponent } from 'react';
 import { DefaultLayout } from '../components/template.js';
 import { RouterComponent } from '../interfaces.js';
-import {
-  LayoutComponent,
-  MDXPageComponent,
-  Metadata,
-  PageComponent,
-  RouteLoaderFunction,
-} from '../types.js';
-import { convertMDXPageToPageComponent, isMDXPage } from './define-router.js';
+import { LayoutComponent, Metadata } from '../types.js';
 
 const basePath = '/src/app/_routes/';
 
-type RouteNode = {
+export type RouteNode = {
   path: string;
   segment: string;
   fullPath: string;
   isLayout?: boolean;
+  module?: () => Promise<Module>;
   component?: FunctionComponent<any>;
   metadata?: Metadata;
-  loader?: RouteLoaderFunction;
   children?: RouteNode[];
+  source?: string;
 };
 
-type Module = {
+export type Module = {
   default: FunctionComponent<any>;
 };
 
@@ -58,7 +52,7 @@ function normalizeSegment(segment: string) {
  * @param foldersOnly Whether to return only folders
  * @returns Path segments
  */
-function getPathSegments(filePath: string, foldersOnly = false) {
+export function getPathSegments(filePath: string, foldersOnly = false) {
   const relative = filePath.replace(basePath, ''); // eg. /src/app/_routes/docs/layout.tsx => docs/layout.tsx
 
   if (!foldersOnly) {
@@ -100,6 +94,7 @@ function generateSkeletonTree(
     segment: '_',
     isLayout: true,
     children: [],
+    source: '',
   };
 
   currentLevel.push(root);
@@ -131,6 +126,7 @@ function generateSkeletonTree(
         segment,
         isLayout: false,
         children: [],
+        source: '',
       };
 
       tmpLevel.push(node);
@@ -151,9 +147,10 @@ function insertNodeToTree(
   // Handle the root layout
   if (segments.length === 1 && segments[0] === '_') {
     currentNode.isLayout = true;
-    currentNode.component = routeInfo.component;
+    currentNode.component = routeInfo.component; // Has to be considered in the case where the developer doesn't provide a layout
     currentNode.metadata = routeInfo.metadata;
-    currentNode.loader = routeInfo.loader;
+    currentNode.module = routeInfo.module;
+    currentNode.source = routeInfo.source;
 
     return;
   }
@@ -184,7 +181,8 @@ function insertNodeToTree(
     currentNode.isLayout = true;
     currentNode.component = routeInfo.component;
     currentNode.metadata = routeInfo.metadata;
-    currentNode.loader = routeInfo.loader;
+    currentNode.module = routeInfo.module;
+    currentNode.source = routeInfo.source;
   } else {
     let path = '';
 
@@ -220,9 +218,8 @@ function insertNodeToTree(
       fullPath: fullPath + '/' + (lastSegment === '.' ? '' : lastSegment),
       segment: lastSegment,
       isLayout: false,
-      component: routeInfo.component,
-      metadata: routeInfo.metadata,
-      loader: routeInfo.loader,
+      module: routeInfo.module,
+      source: routeInfo.source,
     };
 
     currentLevel.push(node);
@@ -240,15 +237,18 @@ async function generateRouter(tree: RouteNode[]) {
   // Generate the base router
   const router = new RouterComponent();
 
+  // use default layout if not defined
+  let layout: RouteNode | LayoutComponent;
+
   // Get layout if defined
   if (root.isLayout) {
-    // use default layout if not defined
-    const layout = (root.component || DefaultLayout) as LayoutComponent;
-    layout.path = root.path || DefaultLayout.path;
-    // TODO: Add metadata here
-
-    router.layout = layout;
-    router.useParentLayout = true;
+    if ('source' in root) {
+      layout = root as RouteNode;
+    } else {
+      layout = root.component as LayoutComponent;
+      layout.path = root.path;
+      layout.metadata = root.metadata;
+    }
   }
 
   // Get pages
@@ -257,6 +257,10 @@ async function generateRouter(tree: RouteNode[]) {
   // Add pages to the router
   router.pages = pages;
   router.routers = routers;
+
+  // Add layout to the router
+  router.layout = layout;
+  router.useParentLayout = true;
 
   return router;
 }
@@ -268,57 +272,37 @@ async function generateRouter(tree: RouteNode[]) {
  */
 async function generateRoutes(tree: RouteNode[]) {
   try {
-    const routes: Array<PageComponent> = [];
+    const routes: Array<RouteNode> = [];
     const routers: RouterComponent[] = [];
 
     for (const node of tree) {
       // Handle page
-      if (!node.isLayout && node.component) {
-        const page = node.component as PageComponent;
-
-        if (!page) {
-          console.warn(
-            `Page component is not exported by default for route: ${node.path}`
-          );
-          continue;
-        }
-
-        if (isMDXPage(page)) {
-          // Convert PageComponent to MDXPageComponent (to make ts happy)
-          const mdxPage = page as unknown as MDXPageComponent;
-
-          mdxPage.metadata.path = node.path;
-          mdxPage.metadata.metadata = node.metadata;
-
-          const pageComponent = await convertMDXPageToPageComponent(mdxPage);
-
-          routes.push(pageComponent);
-          continue;
-        }
-
-        page.path = node.path;
-        page.metadata = node.metadata;
-        page.loader = node.loader;
-
-        routes.push(page);
+      if (!node.isLayout && node.module) {
+        routes.push(node);
 
         continue;
       }
 
       // Handle layout
       if (node.isLayout) {
-        const layout = node.component as LayoutComponent;
+        let layout: LayoutComponent | RouteNode;
 
-        if (!layout) {
-          console.warn(
-            `Layout component is not defined for route: ${node.path}`
-          );
-          continue;
+        if ('source' in node) {
+          layout = node;
+        } else {
+          layout = node.component as LayoutComponent;
+
+          if (!layout) {
+            console.warn(
+              `Layout component is not defined for route: ${node.path}`
+            );
+            continue;
+          }
+
+          layout.path = node.path;
+          layout.metadata = node.metadata;
+          layout.source = node.source;
         }
-
-        layout.path = node.path;
-        layout.metadata = node.metadata;
-        layout.loader = node.loader;
 
         if (node.children) {
           // Loop through children
@@ -366,27 +350,23 @@ async function generateRoutes(tree: RouteNode[]) {
  * @param fn Function that returns a record of modules
  * @returns Router component
  */
-export async function flatRoutes(fn: () => Record<string, Module>) {
+export async function flatRoutes(
+  fn: () => Record<string, () => Promise<Module>>
+) {
   try {
     let modules = fn();
 
-    // import.meta.glob can be undefined in some cases (because it's unavailable outside a vite env)
-    // if (import.meta.glob) {
-    //   let modules = import.meta.glob(
-    //     [
-    //       '/src/app/_routes/**/layout.{jsx,tsx}',
-    //       '/src/app/_routes/**/*.page.{md,mdx,jsx,tsx}',
-    //     ],
-    //     { eager: true }
-    //   );
-    // }
-
     const tree: RouteNode[] = [];
-    const foldersMap: Map<string, { segments: string[]; mod: Module }> =
-      new Map();
-    const modulesMap: Map<string, { segments: string[]; mod: Module }> =
-      new Map();
+    const foldersMap: Map<
+      string,
+      { segments: string[]; mod: () => Promise<Module> }
+    > = new Map();
+    const modulesMap: Map<
+      string,
+      { segments: string[]; mod: () => Promise<Module> }
+    > = new Map();
 
+    // Map the modules and extract segments for folders and modules (layouts and pages)
     for (const [filePath, mod] of Object.entries(modules)) {
       const foldersSegments = getPathSegments(filePath, true);
       const modulesSegments = getPathSegments(filePath);
@@ -405,6 +385,7 @@ export async function flatRoutes(fn: () => Record<string, Module>) {
       )
     );
 
+    // Filter out pages
     const pageModulesMap = new Map(
       [...modulesMap.entries()].filter(([filePath]) =>
         filePath.includes('.page.')
@@ -420,50 +401,21 @@ export async function flatRoutes(fn: () => Record<string, Module>) {
       });
     }
 
+    // Insert every layout into the tree
     for (const [filePath, { segments, mod }] of layoutModulesMap) {
-      if (!mod.default) {
-        console.warn(
-          `Layout component is not exported by default from: ${filePath}}`
-        );
-
-        continue;
-      }
-
-      let metadata = (mod.default as LayoutComponent).metadata;
-      let loader = (mod.default as LayoutComponent).loader;
-
       insertNodeToTree(tree, segments, {
-        component: mod.default,
-        metadata: metadata ?? {},
-        loader,
+        module: mod, // Only present for lazy routes, instead prefer component attribute
         isLayout: true,
+        source: filePath,
       });
     }
 
+    // Insert every pages into the tree
     for (const [filePath, { segments, mod }] of pageModulesMap) {
-      if (!mod.default) {
-        console.warn(
-          `Page component is not exported by default from: ${filePath}}`
-        );
-        continue;
-      }
-
-      let metadata = (mod.default as PageComponent).metadata;
-      let loader = (mod.default as PageComponent).loader;
-
-      // extracting the metadata
-      if (isMDXPage(mod.default)) {
-        metadata = (mod.default as MDXPageComponent).metadata.metadata;
-      }
-
       insertNodeToTree(tree, segments, {
-        component: mod.default,
-        metadata: metadata ?? {
-          title: mod.default.name,
-          description: '',
-        },
-        loader,
+        module: mod, // Only present for lazy routes, instead prefer component attribute
         isLayout: false,
+        source: filePath,
       });
     }