+++ playground/file-based-routing/src/app/app.router.ts
@@ -1,11 +1,12 @@
-// import { RouterComponent, defineRouter } from 'rasengan';
+import { RouterComponent, defineRouter } from 'rasengan';
 import Router from 'virtual:rasengan/router';
+import Page from './test.page';
 
 // class AppRouter extends RouterComponent {}
 
 // export default defineRouter({
 //   imports: [Router],
-//   pages: [],
+//   pages: [Page],
 // })(AppRouter);
 
 export default Router;